Fix CI for item artifact scans and sidebar split
